+++ activerecord/CHANGELOG.md
@@ -1,3 +1,41 @@
+*   Fix nested `has_many :through` associations on unpersisted parent instances.
+
+    For example, if you have
+
+        class Post < ActiveRecord::Base
+          belongs_to :author
+          has_many :books, through: :author
+          has_many :subscriptions, through: :books
+        end
+
+        class Author < ActiveRecord::Base
+          has_one :post
+          has_many :books
+          has_many :subscriptions, through: :books
+        end
+
+        class Book < ActiveRecord::Base
+          belongs_to :author
+          has_many :subscriptions
+        end
+
+        class Subscription < ActiveRecord::Base
+          belongs_to :book
+        end
+
+    Before:
+
+    If `post` is not persisted, then `post.subscriptions` will be empty.
+
+    After:
+
+    If `post` is not persisted, then `post.subscriptions` can be set and used
+    just like it would if `post` were persisted.
+
+    Fixes #16313.
+
+    *Zoltan Kiss*
+
 *   Fixed inconsistency with `first(n)` when used with `limit()`.
     The `first(n)` finder now respects the `limit()`, making it consistent
     with `relation.to_a.first(n)`, and also with the behavior of `last(n)`.